% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digilogger.R
\docType{package}
\name{digilogger-package}
\alias{digilogger-package}
\alias{digilogger}
\title{digilogger - data analysis tools}
\description{
\code{digilogger} package is a toolbox for the analysis of data.
}
\section{Data analysis}{

Some text.
}

\examples{
# Use the NEW function to analyse data
tail(1:100)
}
\author{
Stefan Roediger
}
