% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digilogger_gui.R
\name{digilogger_gui}
\alias{digilogger_gui}
\title{Run digilogger GUI}
\usage{
digilogger_gui()
}
\description{
Run digilogger GUI
}
\details{
This app allows you to upload a single data file and to explore it via several tabs,
one with the summary of measures, one with settings and one with plots.
The app can e.g. be used after one day of field work to quickly check files.
}
\examples{
\dontrun{
digilogger_gui()
}
}
