% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs.import.R
\name{vs.import}
\alias{vs.import}
\title{Import of Data from the VideoScan System}
\usage{
vs.import(file)
}
\arguments{
\item{file}{is the file that contains.}
}
\value{
The output of the vs.import() function is a data.frame
}
\description{
vs.import() is a function to import a file from the VideoScan system.
The \code{\link[utils]{read.csv}} is used internally.
}
\examples{

# Import a csv file from the VideoScan systems
file_name <- "Projekt009383-19-09-2018_FranziD_Chip119_CRP_50_nI_Block.csv"
vs.import(system.file(file_name, package = "digilogger"))

}
\author{
Stefan Roediger, Michal Burdukiewicz
}
\keyword{data}
\keyword{import}
