% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digilogger.R
\docType{package}
\name{digilogger-package}
\alias{digilogger-package}
\alias{digilogger}
\title{digilogger - data analysis tools}
\description{
\code{digilogger} package is a toolbox for the analysis of POCT data.
}
\author{
Stefan Roediger
}
